StatisticalTests

StatisticalTests.TTestModel

StatisticalTests.t_test

StatisticalTests.t_test

StatisticalTests.t_test

StatisticalTests.TTestModel — Type

TTestModel

A TestModel type

Members

X: model matrix

y: vector to compare its means

type: either :one, :paired, :simple, or :welch

unique_levels: String vector of the unique levels (only 1 or 2 elements are allowed)

StatisticalTests.t_test — Function

t_test(X, y, type, unique_levels; μ0=0.0)

An alias for fit(TTestModel, X, y, unique_levels)

The arguments X and y can be a Matrix and a Vector or a Formula and DataFrame.

The argument type must be either :one, :paired, :simple, or welch.

The argument unique_levels can be a Vector{String}.

The keyword argument μ0 can be a Real to use one sample t-test (default = 0.0). Ignored if the type is not :one.

StatisticalTests.t_test — Method

t_test(x1, x2, paired::Bool=false, varequal::Bool=false)

Two sample t-test

The arguments x1 and x2 can be numeric Vector.

Paired t-test will be performed if the keyword argument paired is true. Simple two sample t-test will be performed if the keyword argument varequal is true. Otherwise, Welch two sample t-test will be performed.

StatisticalTests.t_test — Method

t_test(x, μ0::AbstractFloat=0.0)

One sample t-test

The argument x can be numeric Vector.

One sample t-test will be performed to compare between the mean of x1 and μ0.