StatisticalTests

StatisticalTests.FTestModel

StatisticalTests.TTestModel

StatisticalTests.f_test

StatisticalTests.f_test

StatisticalTests.isna

StatisticalTests.ks_test

StatisticalTests.lchoose

StatisticalTests.t_test

StatisticalTests.t_test

StatisticalTests.t_test

StatisticalTests.FTestModel — Type

FTestModel

A TestModel type for f-test

Members

X: model matrix

y: vector to compare its means

levels: String vector of the unique levels (only 2 elements are allowed)

StatisticalTests.TTestModel — Type

TTestModel

A TestModel type for t-test

Members

X: model matrix

y: vector to compare its means

type: either :one, :paired, :simple, or :welch

levels: String vector of the unique levels (only 1 or 2 elements are allowed)

StatisticalTests.f_test — Method

f_test(x1, x2)

F-test

The arguments x1 and x2 can be numeric Vector.

StatisticalTests.f_test — Method

f_test(X, y)

An alias for fit(FTestModel, X, y, levels)

The arguments X and y can be a Matrix and a Vector or a Formula and DataFrame.

The argument levels can be a Vector{String}.

StatisticalTests.isna — Method

isna(x)

Indicate whether x is missing or NaN.

StatisticalTests.ks_test — Method

ks_test(x, distr)

One sample Kolmogorov-Smirnov test

The argument x can be a numeric Vector.

The argument distr can be a Distributions.UnivariateDistribution.

StatisticalTests.lchoose — Method

lchoose(n, j)

Return the log of the number of choose j from n

StatisticalTests.t_test — Function

t_test(X, y, paired::Bool=false, varequal::Bool=false)

An alias for fit(TTestModel, X, y, levels)

The arguments X and y can be a Matrix and a Vector or a Formula and DataFrame.

The argument type must be either :one, :paired, :simple, or welch.

The argument levels can be a Vector{String}.

The keyword argument μ0 can be a Real to use one sample t-test (default = 0.0). Ignored if the type is not :one.

StatisticalTests.t_test — Method

t_test(x1, x2, paired::Bool=false, varequal::Bool=false)

Two sample t-test

The arguments x1 and x2 can be numeric Vector.

Paired t-test will be performed if the keyword argument paired is true. Simple two sample t-test will be performed if the keyword argument varequal is true. Otherwise, Welch two sample t-test will be performed.

StatisticalTests.t_test — Method

t_test(x, μ0::AbstractFloat=0.0)

One sample t-test

The argument x can be numeric Vector.

One sample t-test will be performed to compare between the mean of x1 and μ0.